Calendar foundation: add DB schema, models, DAOs, and service

+++ shale-core/src/main/java/com/shale/core/model/CalendarEvent.java
@@ -7,7 +7,7 @@ public record CalendarEvent(
         int shaleClientId,
         int calendarEventTypeId,
         Integer caseId,
-        Long taskId,
+        Integer taskId,
         String title,
         String description,
         LocalDateTime startsAt,
